Remove AbstractSchedulesResource
